Bundle mir_flutter_app into mir-test-tools snap (#76)
Adds
[`mir-flutter-app`](https://github.com/hbatagelo/mir_flutter_app.git)
built from source into the snap. When the client is executed, it is kept
alive for 120 seconds. This is enough for mir-ci to run the automated
integration tests using the Robot framework.

NB: The reference application needs support for the mir_shell extension.
This is currently not available in the mir_demo_server from the release
ppa.